backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
psycopg[binary]==3.2.6
gunicorn==23.0.0
whitenoise==6.9.0
celery==5.4.0
redis==5.2.1
pandas==2.2.3
scikit-learn==1.6.1
pypdf==5.4.0
